django/pyproject.toml
[tool.poetry]
name = "demsausage"
version = "0.1.0"
description = ""
authors = ["Keith Moss <keithamoss@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.16"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
django-model-utils = "^5.0.0"
googlemaps = "^4.10.0"
gunicorn = "^23.0.0"
gevent = "^24.2.1"
greenlet = "^3.0.0"
social-auth-app-django = "^5.4.2"
social-auth-core = "^4.5.4"
sentry-sdk = "^2.14.0"
django-simple-history = "^3.7.0"
jsonschema = "^4.23.0"
django-pymemcache = "^1.0.0"
djangorestframework-gis = "^1.0"
django-filter = "21.1"
djangorestframework-csv = "^2.1.1"
requests = "^2.31.0"
chardet = "^5.2.0"
drf-spectacular = "^0.27.2"
selenium = "^4.24.0"
pytz = "^2024.2"
rq = "^1.16.2"
django-rq = "^2.10.2"

[tool.poetry.dev-dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-mock = "^3.14"
pytest-xdist = "^3.0"
factory-boy = "^3.3"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "demsausage.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = "--no-migrations -n auto -m 'not slow'"
markers = [
    "slow: marks tests as slow (full-load integration tests; run with -m slow or --run-slow)",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"